# refactor & optimize