--- a/package.json
+++ b/package.json
@@ -27,7 +27,7 @@
   "homepage": "https://github.com/peermaps/kdb-tree-store#readme",
   "devDependencies": {
     "fd-chunk-store": "^2.0.0",
-    "memory-chunk-store": "^1.2.0",
+    "memory-chunk-store": "^1.3.0",
     "tape": "^4.2.2"
   },
   "dependencies": {