This PR applies a great deal of refactoring to the module. I originally tried to plug in a simple in-memory cache for parsed nodes when `_put()` is called, but there was a lot of complex semi-caching happening in the `insert` and `_split[Region|Point]Node` functions. The bulk of the work was refactoring those methods to act more like pure functions that don't modify in-memory state or mutate outside of themselves. Once that work was done, the rest fell into place.

Here are some numbers on speed comparing `master` to this branch:

## Benchmarks

### kdb-tree-store /w memory store

| |  insertion | full map query |
| ---- | ---- | ---- |
| before | 620ms | 15ms |
| after | 202ms |11ms |


| |  insertion | full map query |
| ---- | ---- | ---- |
| before|12555ms|77ms|
| after|2536ms|45ms|

## Digital Democracy datasets (all indexes)

| dataset | before | after |
| ---- | ---- | ---- |
| siekopai | 74043ms | 55199ms
| waorani | 72840ms | 56091ms

Addresses #1.