@noffle 2017-08-15
Merged in 3.2.1.